deleteUser method in Initials component
